requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.4
PyPDF2>=3.0.1
python-docx>=1.0.1
scikit-learn>=1.5.1
pandas>=2.2.2
numpy>=1.26.4
google-cloud>=0.20.0